# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/products/{id}:
    get:
      tags:
      - product-controller
      operationId: getProduct
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductResponse'
    put:
      tags:
      - product-controller
      summary: Update an existing product
      operationId: updateProduct
      parameters:
      - name: id
        in: path
        description: Product ID
        required: true
        schema:
          type: integer
          format: int64
      - name: X-Username
        in: header
        description: Username of the user performing the action
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductResponse'
  /api/master-data/document-types/{code}:
    put:
      tags:
      - document-type-controller
      operationId: updateDocumentType
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentTypeRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
  /api/master-data/document-types/{code}/deactivate:
    put:
      tags:
      - document-type-controller
      operationId: deactivateDocumentType
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
...
  /api/products:
    post:
      tags:
      - product-controller
      summary: Create a new product
      description: Create product with basic info. Requires X-Username header for
        audit.
      operationId: createProduct
      parameters:
      - name: X-Username
        in: header
        description: Username of the user performing the action
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductResponse'
...
  /api/master-data/document-types:
    get:
      tags:
      - document-type-controller
      operationId: getActiveDocumentTypes
      parameters:
      - name: category
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentTypeResponse'
    post:
      tags:
      - document-type-controller
      operationId: createDocumentType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentTypeRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DocumentTypeResponse'
...
  /api/products/{id}/detail:
    get:
      tags:
      - product-controller
      operationId: getProductDetail
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDetailResponse'
  /api/products/search:
    get:
      tags:
      - product-controller
      operationId: searchProducts
      parameters:
      - name: family
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: customerType
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageProductResponse'
  /api/products/code/{code}:
    get:
      tags:
      - product-controller
      operationId: getProductByCode
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductResponse'
...
  /api/limits/assignments/product/{code}:
    get:
      tags:
      - limit-assignment-controller
      operationId: getProductLimits
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductLimitResponse'
...
  /api/charges/interest/product/{productCode}:
    get:
      tags:
      - interest-rate-controller
      operationId: getInterestRatesByProduct
      parameters:
      - name: productCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InterestRateResponse'
  /api/charges/assignments/product/{productCode}:
    get:
      tags:
      - charge-assignment-controller
      operationId: getProductCharges
      parameters:
      - name: productCode
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductChargeResponse'
...
components:
  schemas:
...
    ProductFeeEntryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        feeType:
          type: string
        calculationMethod:
          type: string
        amount:
          type: number
        rate:
          type: number
        currency:
          type: string
        tiers:
          type: array
          items:
            $ref: '#/components/schemas/ProductFeeTierResponse'
    ProductFeeTierResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tierFrom:
          type: integer
          format: int64
        tierTo:
          type: integer
          format: int64
        rate:
          type: number
...
    ProductFeatureResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        featureKey:
          type: string
        featureValue:
          type: string
    UpdateProductRequest:
      type: object
      properties:
        code:
          type: string
        name:
          maxLength: 200
          minLength: 0
          type: string
        description:
          type: string
        customerTypes:
          type: array
          items:
            type: string
            enum:
            - CORPORATE
            - SME
            - INDIVIDUAL
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        family:
          type: string
        status:
          type: string
        versionNumber:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateDocumentTypeRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
    DocumentTypeResponse:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    InterestRateResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        chargeDefinitionId:
          type: integer
          format: int64
        productCode:
          type: string
        fixedRate:
          type: number
        accrualSchedule:
          type: string
        applicationSchedule:
          type: string
        tiers:
          type: array
          items:
            $ref: '#/components/schemas/InterestTierResponse'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InterestTierResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        balanceFrom:
          type: number
        balanceTo:
          type: number
        rate:
          type: number
...
    CreateProductRequest:
      required:
      - code
      - family
      - name
      type: object
      properties:
        code:
          maxLength: 50
          minLength: 0
          type: string
        name:
          maxLength: 200
          minLength: 0
          type: string
        family:
          type: string
          enum:
          - ACCOUNT
          - PAYMENT
          - TRADE_FINANCE
        description:
          type: string
        customerTypes:
          type: array
          items:
            type: string
            enum:
            - CORPORATE
            - SME
            - INDIVIDUAL
...
    CreateDocumentTypeRequest:
      required:
      - category
      - code
      - name
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
...
    ProductLimitResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        limitDefinitionId:
          type: integer
          format: int64
        limitName:
          type: string
        productCode:
          type: string
        overrideAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProductChargeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        chargeDefinitionId:
          type: integer
          format: int64
        chargeName:
          type: string
        productCode:
          type: string
        overrideAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProductDetailResponse:
      type: object
      properties:
        product:
          $ref: '#/components/schemas/ProductInfo'
        id:
          type: integer
          format: int64
        versionNumber:
          type: integer
          format: int64
        status:
          type: string
        contractCount:
          type: integer
          format: int32
        submittedBy:
          type: string
        approvedBy:
          type: string
        rejectionComment:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        features:
          type: array
          items:
            $ref: '#/components/schemas/ProductFeatureResponse'
        feeEntries:
          type: array
          deprecated: true
          items:
            $ref: '#/components/schemas/ProductFeeEntryResponse'
        segments:
          type: array
          items:
            $ref: '#/components/schemas/SegmentInfo'
    ProductInfo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
        family:
          type: string
        description:
          type: string
    SegmentInfo:
      type: object
      properties:
        customerType:
          type: string
    PageProductResponse:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        pageable:
          $ref: '#/components/schemas/PageableObject'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        number:
          type: integer
          format: int32
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          type: array
          items:
            $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        direction:
          type: string
        nullHandling:
          type: string
        ascending:
          type: boolean
        property:
          type: string
        ignoreCase:
          type: boolean
...